cleanup of sketcher
removing try except around update and shift methods...this is not proper.
instead if there is nothing to regenerate I set the generated variable tuple to None and test before updating or constrained_shifting
new constraint-based solving is still slower, I suspect this is due to sympy dependencies.  it is commented out in dq and j for now.